requirements.txt
cdqa==1.3.9
docx2txt==0.8
fpdf==1.7.2
gensim==3.8.3
matplotlib==3.2.2
nltk==3.3
pandas==0.25.0
PyPDF2==1.26.0
PyQt==5.15.0
spacy==2.3.1
torch==1.2.0+cpu